gastronomy_fleet/requirements.txt
google-adk
uvicorn
langchain
langchain-community
langchain-google-vertexai
faiss-cpu
python-dotenv
google-cloud-storage 